+++ src/MarginTrading.BrokerBase/Services/IRabbitMqPoisonQueueHandler.cs
@@ -1,9 +1,0 @@
-using System.Threading.Tasks;
-
-namespace Lykke.MarginTrading.BrokerBase.Services
-{
-    public interface IRabbitMqPoisonQueueHandler
-    {
-        Task<string> PutMessagesBack();
-    }
-}